refactor (moving address validation and normalization into common repository) modified default inserts
