+++ serviceregistry/src/main/java/eu/arrowhead/serviceregistry/service/normalization/DeviceDiscoveryNormalization.java
@@ -10,10 +10,10 @@ import org.springframework.stereotype.Service;
 import org.springframework.util.Assert;
 
 import eu.arrowhead.common.Utilities;
+import eu.arrowhead.common.service.validation.address.AddressNormalizer;
 import eu.arrowhead.dto.AddressDTO;
 import eu.arrowhead.dto.DeviceLookupRequestDTO;
 import eu.arrowhead.dto.DeviceRequestDTO;
-import eu.arrowhead.serviceregistry.service.validation.address.AddressNormalizer;
 
 @Service
 public class DeviceDiscoveryNormalization {
